all: split zip out into subpackage
Building on the discussion in #180, this is a way for callers to be
able to use the zip package directly without importing the other
compression/decompression packages. It uses aliases to preserve
backwards compatibility.

The utility functions and types used by the zip package needed to
be split out into a subpackage as well (named "common") to avoid a
circular dependency issue.