Rework ArchiveAsync interface to make it return any archiving errors
After experience with the current ArchiveAsync interface while
creating an rclone backend it became clear that it wasn't quite right.

With the old interface, callers did not know when the File had been
archived and thus did not know when to release resources associated
with that file.

This patch changes the interface so that it returns the error
archiving each File. This means the caller can know when the File has
been archived and when to release resources. It returns the error for
each file archived which is useful too.

Fixes #368