Allow Identify to take a nil io.Reader
The Match functions can all take a nil io.Reader to mean just identify
the archive from its file name, but before this patch Identify could
not.

This patch allows the stream to be nil, and enables Identify to be
used to create formats for archives which haven't been created yet
just from their file name.